This change fixes the initPage event for post-back requests.  Previously we had a custom UIViewRoot (LayoutViewRoot) which caused the initPage events to fire during decode() (ApplyRequestValuesPhase).  However, we no longer have a custom ViewRoot (for better compatibility with other JSF frameworks), so this functionality was inadvertently lost.
I added back the code to invoke the initPage event at the end of
restoreView(), which is nearly identical to the beginning of the
ApplyRequestValuesPhase and should work exactly as it did before (unless
you're checking the phase id, which I believe is private and nearly
impossible to check... although dynafaces does it, but not via JSFT
handlers).  FYI, the initPage is invoked indirectly via
getLayoutDefinition(), which is why I missed the fact this got removed.

While I was at it, and while I am now getting the LayoutDefinition... I
decided this would be a good place to invoke the decode() handlers as well
as it is also functionally equivalant to its old behavior as it also used
to exist in the UIViewRoot's decode() method.  So page-level "decode"
events now work too!

Enjoy.
GlassFish IT #: 4237, 4226

svn path=/trunk/; revision=1066